refactor(dashboard): extract useSubscribedSnapshot and useSubscribedValue to use-signal-value
Remove redundant hook definitions from ide-shell.ts bottom.
ide-shell.ts now imports these from ./use-signal-value.
