+++ dashboard/src/components/ide/use-signal-value.ts
@@ -10,3 +10,42 @@ export function useSignalValue<T>(
   }, [signal])
   return value
 }
+
+export function useSubscribedValue<T>(
+  read: () => T,
+  subscribe: (listener: () => void) => () => void,
+): T {
+  const [value, setValue] = useState<T>(() => read())
+
+  useEffect(() => {
+    setValue(read())
+    return subscribe(() => {
+      setValue(read())
+    })
+  }, [read, subscribe])
+
+  return value
+}
+
+export function useSubscribedSnapshot<T>(
+  read: () => T,
+  subscribe: (listener: () => void) => () => void,
+): T {
+  const [value, setValue] = useState<T>(() => read())
+
+  useEffect(() => {
+    let current = read()
+    let sawInitialSnapshot = false
+    return subscribe(() => {
+      const next = read()
+      if (!sawInitialSnapshot) {
+        sawInitialSnapshot = true
+        if (next === current) return
+      }
+      current = next
+      setValue(previous => previous === next ? previous : next)
+    })
+  }, [read, subscribe])
+
+  return value
+}
